Better software architecture:
- The method response object is now only responsible for validating
itself.
- The RPC interface (JSON) handles the building the request and parsing
the HTTP response to create a method response object.
