+++ src/citric/rpc/__init__.py
@@ -1,0 +1,1 @@
+"""RPC interfaces."""
