+++ src/citric/session.py
@@ -2,69 +2,9 @@
 from types import TracebackType
 from typing import Any, Optional, Type, TypeVar
 
-import requests
-
 from citric.response import RPCResponse
-
-
-class BaseRPC:
-    """Base class for executing RPC in the LimeSurvey."""
-
-    def invoke(
-        self, url: str, method: str, *args: Any, request_id: int = 1  # noqa: ANN101
-    ) -> RPCResponse:
-        """Execute a LimeSurvey RPC.
-
-        Args:
-            url: URL of LimeSurvey RPC interface.
-            method: Name of the method to call.
-            args: Positional arguments of the RPC method.
-            request_id: Request ID for response validation.
-
-        Raises:
-            NotImplementedError: Subclass does not implement this method.
-        """
-        raise NotImplementedError
-
-
-class JSONRPC(BaseRPC):
-    """Execute JSON-RPC in LimeSurvey."""
-
-    _headers = {
-        "content-type": "application/json",
-        "user-agent": "citric-client",
-    }
-
-    def __init__(self) -> None:  # noqa: ANN101
-        """Create a JSON RPC object."""
-        self.request_session = requests.Session()
-        self.request_session.headers.update(self._headers)
-
-    def invoke(
-        self, url: str, method: str, *args: Any, request_id: int = 1,  # noqa: ANN101
-    ) -> RPCResponse:
-        """Execute a LimeSurvey RPC with a JSON payload.
-
-        Args:
-            url: URL of LimeSurvey RPC interface.
-            method: Name of the method to call.
-            args: Positional arguments of the RPC method.
-            request_id: Request ID for response validation.
-
-        Returns:
-            An RPC response with result, error and id attributes.
-        """
-        payload = {
-            "method": method,
-            "params": [*args],
-            "id": request_id,
-        }
-
-        res = self.request_session.post(url, json=payload)
-
-        response = RPCResponse.parse_response(res)
-
-        return response
+from citric.rpc.base import BaseRPC
+from citric.rpc.json import JSONRPC
 
 
 T = TypeVar("T", bound="Session")
@@ -109,9 +49,12 @@ class Session(object):
         Returns:
             An RPC response with result, error and id attributes.
         """
-        return self.spec.invoke(
+        response = self.spec.invoke(
             self.url, method, self.key, *args, request_id=request_id,
         )
+        response.validate()
+
+        return response
 
     def get_session_key(
         self, admin_user: str, admin_pass: str, request_id: int = 1  # noqa: ANN101
@@ -131,6 +74,7 @@ class Session(object):
         response = self.spec.invoke(
             self.url, "get_session_key", admin_user, admin_pass, request_id=request_id,
         )
+        response.validate()
 
         return response.result
